added new DX and DY output columns to the GRAD line type #118

+++ sql/mv_mysql.sql
@@ -800,8 +800,8 @@ CREATE TABLE line_data_grad
   s1 DOUBLE,
   s1_og DOUBLE DEFAULT -9999,
   fgog_ratio DOUBLE DEFAULT -9999,
-  dx DOUBLE DEFAULT -9999,
-  dy DOUBLE DEFAULT -9999,
+  dx INT DEFAULT -9999,
+  dy INT DEFAULT -9999,
 
   CONSTRAINT line_data_grad_data_file_id_pk
   FOREIGN KEY (data_file_id)
